Added functionality to start SWM components as daemons in background
The script start_swm.sh launched a terminal for every SWM component.
That is useful for testing but hinders CI. Enabled all SWM components
so that they can be started in the foreground as well as daemons in the
background. The start_swm.sh script has a new option '-b' that launches
background mode. The new script stop_swm.sh can be used to stop the
running SWM components.

Signed-off-by: Rudolf J Streif <rudolf.streif@gmail.com>